Fixed warnings

+++ Biemgine/entities/bmSpriteEntity.cpp
@@ -8,7 +8,7 @@ namespace biemgine
     bmSpriteEntity::bmSpriteEntity(std::string texture, float x, float y, bmColor color, float w, float h)
     {
         addComponent("position", new bmPositionComponent(x, y));
-        addComponent("texture", new bmTextureComponent(texture, 0, 0, w, h));
+        addComponent("texture", new bmTextureComponent(texture, 0, 0, static_cast<int>(w), static_cast<int>(h)));
         addComponent("color", new bmColorComponent(color));
     }
 }
